+++ app/Http/Controllers/Backend/WorkoutController.php
@@ -11,6 +11,7 @@ use App\Services\Workout\WorkoutService;
 use App\Services\User\UserService;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Response;
+use App\Policies\Abilities;
 
 class WorkoutController extends Controller
 {
@@ -52,10 +53,12 @@ class WorkoutController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
+        $this->authorize(Abilities::VIEW_ANY, Workout::class);
         return view('backend.pages.workout.index', [
             'workouts' => $this->workoutService->paginate(),
         ]);
@@ -64,10 +67,12 @@ class WorkoutController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize(Abilities::CREATE, Workout::class);
         $users = [
             // @todo Брать значение по умолчанию из конфига
             '' => '– Please select –'
@@ -92,10 +97,12 @@ class WorkoutController extends Controller
      * Store a newly created resource in storage.
      *
      * @param  WorkoutStoreFormRequest  $request
-     * @return Response
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(WorkoutStoreFormRequest $request)
     {
+        $this->authorize(Abilities::CREATE, Workout::class);
         $this->workoutService->create($request->all());
         // @todo Сообщение об успешном создании записи
         return redirect(route('backend.workout.index'));
@@ -116,10 +123,12 @@ class WorkoutController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  Workout  $workout
-     * @return Response
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Workout $workout)
     {
+        $this->authorize(Abilities::UPDATE, $workout);
         $users = [
             // @todo Брать значение по умолчанию из конфига
             '' => '– Please select –'
@@ -146,10 +155,12 @@ class WorkoutController extends Controller
      *
      * @param  WorkoutUpdateFormRequest  $request
      * @param  Workout  $workout
-     * @return Response
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(WorkoutUpdateFormRequest $request, Workout $workout)
     {
+        $this->authorize(Abilities::UPDATE, $workout);
         $this->workoutService->update($workout, $request->all());
         // @todo Сообщение об успешном обновлении записи
         return redirect(route('backend.workout.index'));
@@ -159,10 +170,12 @@ class WorkoutController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  Workout  $workout
-     * @return Response
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Workout $workout)
     {
+        $this->authorize(Abilities::DELETE, $workout);
         // @todo Промежуточная форма подтверждения
         $this->workoutService->delete($workout);
         // @todo Сообщение об успешном удалении записи
